requirements.txt
redis
python-dotenv
pymongo